+++ app/src/main/java/com/example/fitpet/ui/insurance/charge/cause/calendar/CalendarAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.fitpet.ui.insurance.charge.cause
+package com.example.fitpet.ui.insurance.charge.cause.calendar
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
