+++ app/src/main/java/com/example/fitpet/ui/insurance/charge/cause/calendar/CalendarViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.fitpet.ui.insurance.charge.cause
+package com.example.fitpet.ui.insurance.charge.cause.calendar
 
 import androidx.lifecycle.viewModelScope
 import com.example.fitpet.base.BaseViewModel
