+++ app/src/main/res/layout/bottom_sheet_calendar.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.fitpet.ui.insurance.charge.cause.CalendarViewModel" />
+            type="com.example.fitpet.ui.insurance.charge.cause.calendar.CalendarViewModel" />
 
     </data>
