+++ app/src/main/res/layout/fragment_insurance_charge_cause.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.example.fitpet.ui.insurance.charge.InsuranceChargeCauseViewModel" />
+            type="com.example.fitpet.ui.insurance.charge.cause.InsuranceChargeCauseViewModel" />
 
     </data>
